+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Note that since we don't clearly distinguish between a public and private interf
 - Fix units transform data not fully updated when structure child changes
 - Fix `addIndexPairBonds` quadratic runtime case
 - Use adjoint matrix to transform normals in shaders
+- Fix resize handling in `tests/browser`
 
 ## [v4.9.1] - 2024-12-05